Move some test to basic test suite
