+++ tests/tests/neg/builtin_assoc_refts/align_of00.rs
@@ -1,4 +1,4 @@
-use flux_rs::attrs::*;
+use flux_attrs::*;
 
 #[repr(align(4))]
 struct A(bool);
