+++ tests/tests/neg/error_messages/annot_check/fun_sig_error_impl.rs
@@ -1,0 +1,10 @@
+use flux_attrs::*;
+
+trait MyTrait {}
+
+impl MyTrait for i32 {}
+
+#[spec(fn() -> i32)] //~ ERROR function has an incompatible refinement annotation
+fn foo() -> impl MyTrait {
+    0
+}
