+++ tests/tests/neg/error_messages/parse.rs
@@ -1,3 +1,5 @@
+use flux_attrs::*;
+
 #[flux::opaque]
 struct S {
     #[flux::field(i32[0])] //~ ERROR opaque struct can't have refined fields
@@ -9,13 +11,13 @@ enum E {
     A(i32), //~ ERROR missing variant
 }
 
-flux_rs::defs! {
+defs! {
     fn test00() -> bool {
         true <=> false <=> true //~ ERROR syntax error
     }
 }
 
-flux_rs::defs! {
+defs! {
     fn test01() -> bool {
         0 < 1 < 2 //~ ERROR syntax error
     }
