+++ tests/tests/neg/error_messages/wf/bitvec_const01.rs
@@ -11,33 +11,33 @@ impl BV32 {
     }
 }
 
-#[flux_rs::constant(bv_int_to_bv32(0x4567))]
+#[flux::constant(bv_int_to_bv32(0x4567))]
 pub const START: BV32 = BV32::new(0x4567);
 
-#[flux_rs::constant(true)] //~ ERROR mismatched sorts
+#[flux::constant(true)] //~ ERROR mismatched sorts
 pub const BAD1: BV32 = BV32::new(0x4567);
 
-#[flux_rs::constant("hola")] //~ ERROR mismatched sorts
+#[flux::constant("hola")] //~ ERROR mismatched sorts
 pub const BAD2: BV32 = BV32::new(0x4);
 
-#[flux_rs::sig(fn () -> BV32[BAD1])]
+#[flux::sig(fn () -> BV32[BAD1])]
 pub fn test1() -> BV32 {
     BV32::new(0x4567)
 }
 
-#[flux_rs::sig(fn () -> BV32[BAD2])]
+#[flux::sig(fn () -> BV32[BAD2])]
 pub fn test2() -> BV32 {
     BV32::new(0x4567)
 }
 
 pub const BAD3: BV32 = BV32::new(0x4567);
 
-#[flux_rs::sig(fn () -> BV32{v: v == BAD3})] //~ ERROR constant annotation required
+#[flux::sig(fn () -> BV32{v: v == BAD3})] //~ ERROR constant annotation required
 pub fn test3() -> BV32 {
     BV32::new(0x4567)
 }
 
-#[flux_rs::sig(fn () -> BV32{v: v == BAD3})] //~ ERROR constant annotation required
+#[flux::sig(fn () -> BV32{v: v == BAD3})] //~ ERROR constant annotation required
 pub fn test4() -> BV32 {
     BAD3
 }
