--- tests/tests/with_deps/neg/error_messages/annot_check/fun_sig_error_impl.rs
@@ -1,8 +1,0 @@
-trait MyTrait {}
-
-impl MyTrait for i32 {}
-
-#[flux_rs::spec(fn() -> i32)] //~ ERROR function has an incompatible refinement annotation
-fn foo() -> impl MyTrait {
-    0
-}
