--- tests/tests/with_deps/neg/error_messages/wf/missing_constant_annot.rs
@@ -1,12 +1,0 @@
-#[flux_rs::opaque]
-#[flux_rs::refined_by(x: int)]
-pub struct S {
-    x: i32,
-}
-
-const SOME_S: S = S { x: 42 };
-
-#[flux_rs::sig(fn() -> S[SOME_S])] //~ ERROR constant annotation required
-pub fn foo() -> S {
-    SOME_S
-}
